+++ daikon-kafka/kafka-reliable-producer-mongo/src/main/java/org/talend/daikon/kafka/mongo/MongoPendingRecordRepository.java
@@ -1,0 +1,11 @@
+package org.talend.daikon.kafka.mongo;
+
+import java.util.List;
+
+import org.springframework.data.mongodb.repository.MongoRepository;
+import org.talend.daikon.kafka.RecordPriority;
+
+public interface MongoPendingRecordRepository extends MongoRepository<PersistentPendingRecord, String> {
+
+    List<PersistentPendingRecord> findByApplicationNameAndRecordPriority(String applicationName, RecordPriority priority);
+}
