+++ daikon-kafka/kafka-reliable-producer-mongo/src/test/resources/logback.xml
