+++ daikon-kafka/kafka-reliable-producer-spring/src/main/java/org/talend/daikon/kafka/KafkaReliableProducerConfiguration.java
@@ -1,0 +1,14 @@
+package org.talend.daikon.kafka;
+
+import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
+import org.springframework.context.annotation.ComponentScan;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.scheduling.annotation.EnableScheduling;
+
+@ComponentScan(basePackages = { "org.talend.daikon.kafka" })
+@Configuration
+@EnableScheduling
+@EnableAutoConfiguration
+public class KafkaReliableProducerConfiguration {
+
+}
