+++ daikon-kafka/kafka-reliable-producer-spring/src/test/resources/logback.xml
@@ -1,0 +1,28 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!--
+  ~ ============================================================================
+  ~
+  ~ Copyright (C) 2006-2015 Talend Inc. - www.talend.com
+  ~
+  ~ This source code is available under agreement available at
+  ~ %InstallDIR%\features\org.talend.rcp.branding.%PRODUCTNAME%\%PRODUCTNAME%license.txt
+  ~
+  ~ You should have received a copy of the agreement
+  ~ along with this program; if not, write to Talend SA
+  ~ 9 rue Pages 92150 Suresnes, France
+  ~
+  ~ ============================================================================
+  -->
+
+<configuration>
+  <appender name="CONSOLE" class="ch.qos.logback.core.ConsoleAppender">
+    <encoder>
+      <pattern>%d %5p | %t | %-55logger{55} | %m %n</pattern>
+    </encoder>
+  </appender>
+
+  <root>
+    <level value="WARN" />
+    <appender-ref ref="CONSOLE" />
+  </root>
+</configuration>
