+++ daikon-kafka/kafka-reliable-producer/src/test/resources/log4j.properties
