--- daikon-kafka/src/main/java/org/talend/daikon/DaikonKafkaConfiguration.java
@@ -1,13 +1,0 @@
-package org.talend.daikon;
-
-import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
-import org.springframework.context.annotation.ComponentScan;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.scheduling.annotation.EnableScheduling;
-
-@ComponentScan(basePackages = { "org.talend.daikon.kafka" })
-@Configuration
-@EnableScheduling
-@EnableAutoConfiguration
-public class DaikonKafkaConfiguration {
-}
